CD-48: поправил запись в таблицу с историей изменения

+++ Analys_prostoev/MainWindow.xaml.cs
@@ -10,6 +10,7 @@ using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
 using System.Windows.Forms;
 using System.Windows.Input;
+using MessageBox = System.Windows.MessageBox;
 
 namespace Analys_prostoev
 {
@@ -82,15 +83,23 @@ namespace Analys_prostoev
                 connection.Open();
 
                 DataRowView item = (DataRowView)DataGridTable.SelectedItem;
-                long id = (long)item.Row["Id"]; ;
-                string deleteQuery = $"DELETE FROM analysis WHERE \"Id\" = {id}";
-
-                using (NpgsqlCommand deleteCommand = new NpgsqlCommand(deleteQuery, connection))
+                if (item == null)
                 {
-                    int rowsAffected = deleteCommand.ExecuteNonQuery();
-                    if (rowsAffected > 0)
+                    MessageBox.Show("Вы не выбрали простой!");
+                    return;
+                }
+                else
+                {
+                    long id = (long)item.Row["Id"]; ;
+                    string deleteQuery = $"DELETE FROM analysis WHERE \"Id\" = {id}";
+
+                    using (NpgsqlCommand deleteCommand = new NpgsqlCommand(deleteQuery, connection))
                     {
-                        System.Windows.MessageBox.Show("Запись удалена успешно.\nСделайте повторную загрузку для обновления таблицы");
+                        int rowsAffected = deleteCommand.ExecuteNonQuery();
+                        if (rowsAffected > 0)
+                        {
+                            System.Windows.MessageBox.Show("Запись удалена успешно.\nСделайте повторную загрузку для обновления таблицы");
+                        }
                     }
                 }
             }
@@ -109,11 +118,19 @@ namespace Analys_prostoev
         private void ChangeMenuItem_Click(object sender, RoutedEventArgs e)
         {
             DataRowView item = (DataRowView)DataGridTable.SelectedItem;
-            ChangeTimeDown change = new ChangeTimeDown(Convert.ToDateTime(item.Row.ItemArray[1]),
-                Convert.ToDateTime(item.Row.ItemArray[2]), Convert.ToInt32(item.Row.ItemArray[3]),
-                Convert.ToString(item.Row.ItemArray[4]), Convert.ToInt64(item.Row.ItemArray[0]));
+            if (item == null)
+            {
+                MessageBox.Show("Вы не выбрали простой!");
+                return;
+            }
+            else
+            {
+                ChangeTimeDown change = new ChangeTimeDown(Convert.ToDateTime(item.Row.ItemArray[1]),
+                               Convert.ToDateTime(item.Row.ItemArray[2]), Convert.ToInt32(item.Row.ItemArray[3]),
+                               Convert.ToString(item.Row.ItemArray[4]), Convert.ToInt64(item.Row.ItemArray[0]));
 
-            change.Show();
+                change.Show();
+            }
         }
 
         private void Button_Click(object sender, RoutedEventArgs e)
